+++ index.js
@@ -6,11 +6,13 @@ const debug = createDebug('puppeteer-loadtest');
 const exec = require('child_process').exec;
 const argv = require('minimist')(process.argv.slice(2));
 const perf = require('execution-time')();
+const fs = require('fs');
 
 const file = argv.file;
 const samplesRequested = argv.s || 1;
 const concurrencyRequested = argv.c || 1;
 const silent = argv.silent || false;
+const outputFile = argv.outputFile;
 
 if (!file) {
   return console.error('cannot find --file option');
@@ -33,14 +35,32 @@ debug('puppeteer-loadtest is loading...');
 const cmdToExecute = `node ${file}`;
 const results = {};
 let samplesCount = 0;
-let concurrencyCount = 0;
 
-const executeTheCommand = function(cmd) {
+const startSampleLogPerformance = () => {
+  perf.start(`sampleCall${samplesCount + 1}`);
+  results[`sample${samplesCount + 1}`] = {};
+};
+
+const stopSampleLogPerformance = () => {
+  results[`sample${samplesCount + 1}`].sample = perf.stop(`sampleCall${samplesCount + 1}`);
+};
+
+const startConcurrencyLogPerformance = (concurrencyCount) => {
+  perf.start(`sample${samplesCount + 1}concurrencyCount${concurrencyCount + 1}`);
+  results[`sample${samplesCount + 1}`].concurrency = {};
+}
+
+const stopConcurrencyLogPerformance = (concurrencyCount) => {
+  results[`sample${samplesCount + 1}`].concurrency[`${concurrencyCount + 1}`] = perf.stop((`sample${samplesCount + 1}concurrencyCount${concurrencyCount + 1}`));
+}
+
+
+const executeTheCommand = function(cmd, concurrencyCount) {
   return new Promise((resolve, reject) => {
+    startConcurrencyLogPerformance(concurrencyCount);
     exec(cmd, function(error, stdout, stderr) {
-      concurrencyCount += 1;
-      debug(`sample: ${samplesCount}`);
-      debug(`concurrent: ${concurrencyCount}`);
+      debug(`sample: ${samplesCount}, concurrent: ${concurrencyCount}`);
+      stopConcurrencyLogPerformance(concurrencyCount);
       if(stderr) reject(stderr);
       if(error) reject(error);
       resolve(stdout);
@@ -48,32 +68,40 @@ const executeTheCommand = function(cmd) {
   });   
 };
 
-const doAnotherSample = () => { 
+const doAnotherSample = async () => { 
   if(samplesCount < samplesRequested) {
-    doConcurrency();
+    startSampleLogPerformance();
+    await doConcurrency(results);
+    stopSampleLogPerformance();
     samplesCount += 1;
+    return doAnotherSample();
+  }
+  if (outputFile) {
+    fs.writeFileSync(outputFile, JSON.stringify(results, null, "\t"));
+  }
+  if (!silent) {
+    console.log(JSON.stringify(results, null, "\t"));
   }
 };
 
-const doConcurrency = () => {
+const doConcurrency = async (results) => {
   const promisesArray = [];
 
-  for(let i=0; i<concurrencyRequested; i += 1) {
+  for(let i=0; i < concurrencyRequested; i += 1) {
     promisesArray.push(
-      executeTheCommand(cmdToExecute)
+      executeTheCommand(cmdToExecute, i, results)
     );
   }
 
-  concurrencyCount = 0;
-  Promise.all(promisesArray)
-  .then(values => { 
+  let values;
+  try {
+    perf.start('concurrencyCall');
+    values = await Promise.all(promisesArray);
     debug(values);
-    doAnotherSample();
-  })
-  .catch(error => {
+  } catch(error) {
     debug(error);
-    doAnotherSample();
-  });
+  }
+  return values;
 };
 
 doAnotherSample();