+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## Next version
+- Add option (`outputFile`) for measuring performance.
 - update pacakges.
 
 ## 1.0.6